# backend/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: RefactorForge API
  description: |
    RefactorForge is an advanced code intelligence platform that provides comprehensive repository analysis, 
    pattern extraction, and automated refactoring recommendations. The API offers real-time GitHub scanning, 
    performance monitoring, and intelligent code improvement suggestions.

    ## Key Features
    - **Multi-Repository Analysis**: Simultaneous scanning and analysis of multiple GitHub repositories
    - **Pattern Recognition**: Advanced pattern extraction from TypeScript, JavaScript, Python, and other languages
    - **Security Analysis**: Automated detection of security vulnerabilities and misconfigurations
    - **Performance Optimization**: Identification of performance bottlenecks and optimization opportunities
    - **Type Safety Analysis**: Detection of type safety issues and recommendations for improvements
    - **Real-Time Monitoring**: Live GitHub webhook integration for continuous analysis
    - **Recommendation Engine**: AI-powered suggestions for code improvements and refactoring

    ## Authentication
    Most endpoints require authentication via GitHub token for accessing private repositories and higher rate limits.

    ## Rate Limiting
    - **Authenticated**: 5000 requests/hour
    - **Unauthenticated**: 60 requests/hour

    ## Performance Features
    - Connection pooling and request batching
    - Intelligent caching with configurable TTL
...
  version: "2.1.0"
  contact:
...
servers:
...
security:
  - GitHubToken: []
paths:
...
  /api/repositories/{id}/patterns:
    get:
      summary: Get repository patterns
      description: |
        Retrieves all extracted patterns from a specific repository with optional
        filtering and sorting capabilities.
      operationId: getRepositoryPatterns
      tags:
        - Repositories
        - Patterns
      parameters:
        - name: id
          in: path
          description: Repository ID
          required: true
          schema:
            type: string
        - name: category
          in: query
          description: Filter by pattern category
          required: false
          schema:
            type: string
            enum: [function, type, import, react, security, async, error_handling]
        - name: language
          in: query
          description: Filter by programming language
          required: false
          schema:
            type: string
        - name: confidence
          in: query
          description: Minimum confidence score (0-1)
          required: false
          schema:
            type: number
            minimum: 0
            maximum: 1
      responses:
        '200':
          description: Patterns retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  patterns:
                    type: array
                    items:
                      $ref: '#/components/schemas/Pattern'
                  total:
                    type: integer
                  repository:
                    $ref: '#/components/schemas/Repository'
  /api/patterns:
    get:
      summary: Search patterns across repositories
      description: |
        Searches for patterns across all analyzed repositories with advanced
        filtering, sorting, and aggregation options.
      operationId: searchPatterns
      tags:
        - Patterns
      parameters:
        - name: search
          in: query
          description: Search query for pattern content
          required: false
          schema:
            type: string
        - name: category
          in: query
          description: Filter by pattern category
          required: false
          schema:
            type: string
        - name: language
          in: query
          description: Filter by programming language
          required: false
          schema:
            type: string
        - name: repository
          in: query
          description: Filter by repository name
          required: false
          schema:
            type: string
        - name: page
          in: query
          description: Page number for pagination
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of patterns per page
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Patterns retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  patterns:
                    type: array
                    items:
                      $ref: '#/components/schemas/Pattern'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
                  aggregations:
                    type: object
                    properties:
                      byCategory:
                        type: object
                        additionalProperties:
                          type: integer
                      byLanguage:
                        type: object
                        additionalProperties:
                          type: integer
                      byRepository:
                        type: object
                        additionalProperties:
                          type: integer
...
  /api/performance/alerts:
    get:
      summary: Get performance alerts
      description: |
        Retrieves active and recent performance alerts with severity levels
        and resolution status.
      operationId: getPerformanceAlerts
      tags:
        - Performance
      parameters:
        - name: severity
          in: query
          description: Filter by alert severity
          required: false
          schema:
            type: string
            enum: [low, medium, high, critical]
        - name: resolved
          in: query
          description: Filter by resolution status
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: Performance alerts retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  alerts:
                    type: array
                    items:
                      $ref: '#/components/schemas/PerformanceAlert'

    post:
      summary: Resolve performance alert
      description: |
        Manually resolves an active performance alert by ID.
      operationId: resolvePerformanceAlert
      tags:
        - Performance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                alertId:
                  type: string
                  description: Alert ID to resolve
                resolution:
                  type: string
                  description: Optional resolution note
      responses:
        '200':
          description: Alert resolved successfully
        '404':
          description: Alert not found
components:
  securitySchemes:
    GitHubToken:
      type: http
      scheme: bearer
      bearerFormat: GitHub Personal Access Token
      description: |
        GitHub Personal Access Token for accessing private repositories and higher rate limits.
        Generate at: https://github.com/settings/tokens
  schemas:
...
    Repository:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        fullName:
          type: string
        description:
          type: string
        language:
          type: string
        patternsCount:
          type: integer
        lastUpdated:
          type: string
          format: date-time
        lastAnalyzed:
          type: string
          format: date-time
        categories:
          type: array
          items:
            type: string
        techStack:
          type: array
          items:
            type: string
        metrics:
          type: object
          properties:
            linesOfCode:
              type: integer
            fileCount:
              type: integer
            complexity:
              type: number
...
    Pattern:
      type: object
      properties:
        id:
          type: string
        repositoryId:
          type: string
        patternType:
          type: string
        patternName:
          type: string
        content:
          type: string
          description: Code snippet or pattern content
        description:
          type: string
        category:
          type: string
        subcategory:
          type: string
        filePath:
          type: string
        lineStart:
          type: integer
        lineEnd:
          type: integer
        language:
          type: string
        framework:
          type: string
        confidenceScore:
          type: number
          minimum: 0
          maximum: 1
        tags:
          type: array
          items:
            type: string
        contextBefore:
          type: string
        contextAfter:
          type: string
        metadata:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
...
    PerformanceAlert:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [memory, cpu, response_time, error_rate, throughput]
        severity:
          type: string
          enum: [low, medium, high, critical]
        message:
          type: string
        value:
          type: number
        threshold:
          type: number
        timestamp:
          type: string
          format: date-time
        resolved:
          type: boolean
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        pages:
          type: integer
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
...